+++ static_lib/subtract_numbers.c
@@ -1,4 +1,4 @@
-int subtract_numbers (long int x, long int y)
+int subtract_numbers (long long x, long long y)
 {
     return x - y;
 }
